Seed random generator before rendering of each component
